Test folder is extracted from src folder
