+++ tests/integration/controllers/images/photo.jpg
